Remove custom implementation of Ancient Apparition's Ice Vortex
While it demonstrates that fixing the lingering vision issue is
feasible, it is not a general solution and would require that similar
implementations be made for all abilities with problematic temporary
vision, which would add a large maintenance burden.  Therefore, to keep
things both maintainable and consistent, this single implementation is
removed.

(The lingering vision issue is the fact that temporary vision created by
abilities cannot be removed between rounds, necessitating a custom
solution that provides such capability.)